pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bbml"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "torch",
    "torchvision",
    "accelerate",
    "einops",
    "pillow",
    "safetensors",
    "transformers",
    "diffusers",
    "numpy",
    "pydantic",
    "peft",
    "wandb",
    "python-dotenv",
    "torchao",
    "PyYAML",
    "matplotlib",
    "pandas",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["bbml*"]
namespaces = true
